HAL rework to rev2, ADC thermal sampling, minor fixing.

+++ src/pmc.h
@@ -259,8 +259,8 @@ typedef struct {
 pmc_t;
 
 void pmc_default(pmc_t *pm);
-void pmc_feedback(pmc_t *pm, int xA, int xB);
-void pmc_voltage(pmc_t *pm, int xU);
+void pmc_feedback(pmc_t *pm, float xA, float xB);
+void pmc_voltage(pmc_t *pm, float xU);
 void pmc_request(pmc_t *pm, int req);
 
 const char *pmc_strerror(int errno);
